Default to OpenSSL and make BoringSSL optional
Set TLS_BORINGSSL=1 env to link against BoringSSL

+++ mingw/build/_dl.sh
@@ -4,6 +4,7 @@ export MEMCACHED_VER_='1.6.3'
 export MC_CRUSHER_VER_='master'
 export LIBEVENT_VER_='2.1.11-stable'
 export LIBEVENT_HASH=a65bac6202ea8c5609fd5c7e480e6d25de467ea1917c08290c521752f147283d
+export OPENSSL_VER_='1.1.1f'
 export OSSLSIGNCODE_VER_='1.7.1'
 export OSSLSIGNCODE_HASH=f9a8cdb38b9c309326764ebc937cba1523a3a751a7ab05df3ecc99d18ae466c9
 
@@ -56,12 +57,14 @@ rm -f -r libevent && mv libevent-* libevent
 
 # BoringSSL version to be used is the %Y%m%d%H%M date format of the
 # https://github.com/google/boringssl/tree/chromium-stable's latest commit
-rm -rf boringssl
-git clone --branch chromium-stable --depth=1 https://github.com/google/boringssl.git
-[ -f "boringssl${_patsuf}.patch" ] && dos2unix < "boringssl${_patsuf}.patch" | patch --batch -N -p1 -d boringssl
-cd boringssl
-export BORINGSSL_VER_="$(git log --date=format:'%Y%m%d%H%M' -1 | sed '3q;d' | awk -F ' ' '{print $2}')"
-cd ..
+if [ -n "${TLS_BORINGSSL}" ]; then
+    rm -rf boringssl
+    git clone --branch chromium-stable --depth=1 https://github.com/google/boringssl.git
+    [ -f "boringssl${_patsuf}.patch" ] && dos2unix < "boringssl${_patsuf}.patch" | patch --batch -N -p1 -d boringssl
+    cd boringssl
+    export BORINGSSL_VER_="$(git log --date=format:'%Y%m%d%H%M' -1 | sed '3q;d' | awk -F ' ' '{print $2}')"
+    cd ..
+fi
 
 # Official memcached to be used in timestamping since it has no Changelog that can be used as reference
 UPSTREAM_DIR="memcached-${MEMCACHED_VER_}"
